Merge from Portfolio version
Alot of updates were made on the portfolio repo. This is an up-to-date fully functioning version of the project. This should now be a sufficient enough seperation to continue development.

As I train in web development, I will be more careful with how I set up my projects and repositories.

Further information on the javascript implementation of this project can be found on the Portfolio Repository.